HHT-393: renaming transformers
