+++ app/k8s_transform/__fixture__/master_node.json
