+++ app/k8s_transform/__fixture__/replicaset.json
