+++ app/k8s_transform/__fixture__/statefulset.json
