+++ app/k8s_transform/__fixture__/worker_node.turtle
