+++ app/k8s_transform/pod_transformer.py
@@ -4,8 +4,8 @@ import re
 
 from jsonpath_ng.ext import parse
 
+from app.k8s_transform.transformer_base import TransformerBase
 from app.kg.knowledge_graph import KnowledgeGraph, RelationSet
-from app.rdf_transform.transformer_base import TransformerBase
 
 
 class PodToRDFTransformer(TransformerBase):
