+++ app/k8s_transform/test_pod_transformer.py
@@ -1,17 +1,19 @@
 from io import StringIO
 
+from app.k8s_transform.pod_transformer import PodToRDFTransformer
+from app.k8s_transform.test_base import TransformBaseTest
 from app.kg.inmemory_knowledge_graph import InMemoryKnowledgeGraph
 from app.kg.turtle_serializer import TurtleSerialializer
-from app.rdf_transform.replicaset_transformer import ReplicaSetToRDFTransformer
-from app.rdf_transform.test_base import TransformBaseTest
 
 
-class ReplicaSetTransformerTest(TransformBaseTest):
+class PodTransformerTest(TransformBaseTest):
     def setUp(self):
         self.maxDiff = None
 
     def test_transform(self):
-        self.transform("replicaset")
+        self.transform("pod1")
+        self.transform("pod2")
+        self.transform("pod3")
 
     def transform(self, file_id: str) -> None:
         node_json = self.load_json(file_id)
@@ -19,6 +21,6 @@ class ReplicaSetTransformerTest(TransformBaseTest):
 
         buffer = StringIO()
         graph = InMemoryKnowledgeGraph()
-        ReplicaSetToRDFTransformer(node_json, graph).transform()
+        PodToRDFTransformer(node_json, graph).transform()
         TurtleSerialializer().write(buffer, graph)
         self.assertEqual(buffer.getvalue(), node_turtle)
