+++ app/k8s_transform/test_replicaset_transformer.py
@@ -1,19 +1,17 @@
 from io import StringIO
 
+from app.k8s_transform.replicaset_transformer import ReplicaSetToRDFTransformer
+from app.k8s_transform.test_base import TransformBaseTest
 from app.kg.inmemory_knowledge_graph import InMemoryKnowledgeGraph
 from app.kg.turtle_serializer import TurtleSerialializer
-from app.rdf_transform.pod_transformer import PodToRDFTransformer
-from app.rdf_transform.test_base import TransformBaseTest
 
 
-class PodTransformerTest(TransformBaseTest):
+class ReplicaSetTransformerTest(TransformBaseTest):
     def setUp(self):
         self.maxDiff = None
 
     def test_transform(self):
-        self.transform("pod1")
-        self.transform("pod2")
-        self.transform("pod3")
+        self.transform("replicaset")
 
     def transform(self, file_id: str) -> None:
         node_json = self.load_json(file_id)
@@ -21,6 +19,6 @@ class PodTransformerTest(TransformBaseTest):
 
         buffer = StringIO()
         graph = InMemoryKnowledgeGraph()
-        PodToRDFTransformer(node_json, graph).transform()
+        ReplicaSetToRDFTransformer(node_json, graph).transform()
         TurtleSerialializer().write(buffer, graph)
         self.assertEqual(buffer.getvalue(), node_turtle)
