+++ app/k8s_transform/test_workload_transformer.py
@@ -1,9 +1,9 @@
 from io import StringIO
 
+from app.k8s_transform.test_base import TransformBaseTest
+from app.k8s_transform.workload_transformer import WorkloadToRDFTransformer
 from app.kg.inmemory_knowledge_graph import InMemoryKnowledgeGraph
 from app.kg.turtle_serializer import TurtleSerialializer
-from app.rdf_transform.test_base import TransformBaseTest
-from app.rdf_transform.workload_transformer import WorkloadToRDFTransformer
 
 
 class WorkloadTransformerTest(TransformBaseTest):
